\name{Biocarta.nullset}
\docType{data}
\alias{Biocarta.nullset}
\title{Biocarta background set}
\description{This dataset provides the background gene set for Biocarta concepts.}
\usage{Biocarta.nullset}
%- maybe also 'usage' for other objects documented here.
\format{A data frame of length 1254.}
\source{ Biocarta pathway database}
\references{ }
\keyword{datasets }

