\name{DrugBank.annot}
\docType{data}
\alias{DrugBank.annot}
\title{DrugBank annotation}
\description{This dataset links DrugBank concept IDs to their descriptive names.
}
\usage{DrugBank.annot}
%- maybe also 'usage' for other objects documented here.
\format{A data frame of dimensions 256 x 2.}
\source{DrugBank database}
\references{Wishart, D.S., Knox, C., Guo, A.C., Cheng, D., Shrivastava, S., Tzur, D., Gautam, B. and Hassanali, M. (2008) DrugBank: a knowledgebase for drugs, drug actions and drug targets, Nucleic Acids Res, 36, D901-906. }
\keyword{datasets }

