\name{DrugBank}
\docType{data}
\alias{DrugBank}
\title{Human gene sets from targeted by a common entry in DrugBank}
\description{This dataset is a list of DrugBank concept IDs containing the human Entrez Gene IDs assigned to each concept.
}
\usage{DrugBank}
%- maybe also 'usage' for other objects documented here.
\format{A list of length 256.}
\source{DrugBank }
\references{Wishart, D.S., Knox, C., Guo, A.C., Cheng, D., Shrivastava, S., Tzur, D., Gautam, B. and Hassanali, M. (2008) DrugBank: a knowledgebase for drugs, drug actions and drug targets, Nucleic Acids Res, 36, D901-906. }
\keyword{datasets }

