\name{MeSH.nullset}
\docType{data}
\alias{MeSH.nullset}
\title{MeSH background set}
\description{This dataset provides the background gene set for MeSH concepts.}
\usage{MeSH.nullset}
%- maybe also 'usage' for other objects documented here.
\format{A data frame of dimensions 36393 x 1.}
\source{MeSH and gene2mesh}
\references{ }
\keyword{datasets }

