\name{MeSH}
\docType{data}
\alias{MeSH}
\title{Human gene sets linked to common MeSH terms according to Gene2mesh (http://gene2mesh.ncibi.org).}
\description{This dataset is a list of MeSH concept IDs containing the human Entrez Gene IDs assigned to each concept.
}
\usage{MeSH}
%- maybe also 'usage' for other objects documented here.
\format{A list of length 4028.}
\source{MeSH and gene2mesh (http://gene2mesh.ncibi.org)}
\references{ }
\keyword{datasets }

