\name{MiMI.nullset}
\docType{data}
\alias{MiMI.nullset}
\title{MiMI background set}
\description{This dataset provides the background gene set for MiMI concepts.}
\usage{MiMI.nullset}
%- maybe also 'usage' for other objects documented here.
\format{A data frame of dimensions 12447 x 1.}
\source{MiMI database }
\references{ }
\keyword{datasets }

