\name{OMIM.nullset}
\docType{data}
\alias{OMIM.nullset}
\title{OMIM background set}
\description{This dataset provides the background gene set for OMIM concepts.}
\usage{OMIM.nullset}
%- maybe also 'usage' for other objects documented here.
\format{A data frame of dimensions 36393 x 1.}
\source{OMIM database }
\references{ }
\keyword{datasets }

