\name{OMIM}
\docType{data}
\alias{OMIM}
\title{Human gene sets assigned to a common disease in OMIM (Online Mendelian Inheritance in Man).}
\description{This dataset is a list of OMIM concept IDs containing the human Entrez Gene IDs assigned to each concept.
}
\usage{OMIM}
%- maybe also 'usage' for other objects documented here.
\format{A list of length 7050.}
\source{MiMI database (http://mimi.ncibi.org)}
\references{Tarcea, V.G., Weymouth, T., Ade, A., Bookvich, A., Gao, J., Mahavisno, V., Wright, Z., Chapman, A., Jayapandian, M., Ozgur, A., Tian, Y., Cavalcoli, J., Mirel, B., Patel, J., Radev, D., Athey, B., States, D. and Jagadish, H.V. (2008) Michigan molecular interactions r2: from interacting proteins to pathways, Nucleic Acids Res.}
\keyword{datasets }

