\name{Panther.annot}
\docType{data}
\alias{Panther.annot}
\title{Panther annotation}
\description{This dataset links Panther concept IDs to their descriptive names.
}
\usage{Panther.annot}
%- maybe also 'usage' for other objects documented here.
\format{A data frame of dimensions 86 x 2.}
\source{Panther database }
\references{}
\keyword{datasets }

