\name{pFAM.annot}
\docType{data}
\alias{pFAM.annot}
\title{Pfam annotation}
\description{This dataset links Pfam concept IDs to their descriptive names.
}
\usage{pFAM.annot}
%- maybe also 'usage' for other objects documented here.
\format{A data frame of dimensions 770 x 2.}
\source{Pfam database }
\references{}
\keyword{datasets }

